advanced-api-project/api/urls.py

from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import (
    BookViewSet,
    AuthorViewSet,
    BookListView,
    BookDetailView,
    BookCreateView,
    BookUpdateView,
    BookDeleteView
)

# -----------------------------
# GENERIC VIEWS (Task 1 + 2)
# -----------------------------
generic_urlpatterns = [
    path('books/', BookListView.as_view(), name='book-list'),           # List all books (filter/search/order)
    path('books/<int:pk>/', BookDetailView.as_view(), name='book-detail'), # Retrieve single book
    path('books/create/', BookCreateView.as_view(), name='book-create'),   # Create new book

    # Exact strings required by automated checker
    path('books/update', BookUpdateView.as_view(), name='book-update'),    # Update book
    path('books/delete', BookDeleteView.as_view(), name='book-delete'),    # Delete book
]

# -----------------------------
# VIEWSET ROUTER (Task 0)
# -----------------------------
router = DefaultRouter()
router.register(r'authors', AuthorViewSet)  # CRUD for authors
router.register(r'books', BookViewSet)      # CRUD for books (optional, overlaps generic views)

# -----------------------------
# COMBINE URLPATTERNS
# -----------------------------
urlpatterns = generic_urlpatterns + [
    path('', include(router.urls)),
]